Move events sub-components to folder
