--- src/panel/pages/events/constants.ts
@@ -1,3 +1,0 @@
-// Breakpoints
-export const smMax = "399px";
-export const mdMin = "400px";
